chore: clarify download return type, widen semitones to float, update .gitignore

+++ dlt/pitcher.py
@@ -1,5 +1,5 @@
 from pathlib import Path
 
 
-def pitch_shift(wav_path: Path, semitones: int, output_path: Path) -> Path:
+def pitch_shift(wav_path: Path, semitones: float, output_path: Path) -> Path:
     pass
